+++ CHANGELOG.md
@@ -5,4 +5,8 @@ This file is used to list changes made in each version of the oraclexe cookbook.
 
 0.1.0
 -----
-- [frankwis] - Initial release of oraclexe
+- [frankwis | 2015-11-02] - Initial release of oraclexe
+
+0.1.1
+-----
+- [frankwis | 2015-11-03] - Switching to [`magic_shell`](https://github.com/customink-webops/magic_shell) for environment variables